backend/requirements.txt
fastapi
uvicorn
tensorflow
python-dotenv
langchain
langchain-openai
python-multipart